Removed option to create one further course after creating a course.
It turned out that teachers created a new course, used a general template and did not take a look into the newly created course, but directly moved on to create a second course.
To make teachers aware that a there is some work to do in a newly created course, we deleted the option to jump to start a new course request after one course is created. Thus teachers who follow the process finally see the course page.